Fix black binary again for debug python (#86275)
The `--no-binary` flag was not ported when moving from black only to ufmt.
This adds it back.

This is to work around the fact that black binary hard crashes when running with debug python and it needs to be compiled from source.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/86275
Approved by: https://github.com/bdhirsh, https://github.com/malfet